GeneAI/requirements.txt
pdfplumber>=0.11.8
openai>=2.8.0
python-dotenv>=1.2.0
numpy>=2.0.0
pytesseract>=0.3.13
pillow>=12.0.0
scikit-learn>=1.0.0
streamlit>=1.41.0
requests>=2.32.3
flask>=3.0.0
flask-cors>=4.0.0




